inventory_api/requirements.txt
asgiref==3.6.0
black==22.12.0
click==8.1.3
Django==4.1.4
djangorestframework==3.14.0
isort==5.11.4
mypy-extensions==0.4.3
pathspec==0.10.3
platformdirs==2.6.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-decouple==3.6
pytz==2022.7
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
